Ran format, organise imports and cleanup to get everything in sync.

+++ common/crazypants/vecmath/Vector3f.java
@@ -1,7 +1,7 @@
 package crazypants.vecmath;
 
 public class Vector3f {
-  
+
   public float x;
   public float y;
   public float z;
@@ -11,23 +11,23 @@ public class Vector3f {
     y = 0;
     z = 0;
   }
-  
+
   public Vector3f(double x, double y, double z) {
-    this.x = (float)x;
-    this.y = (float)y;
-    this.z = (float)z;
+    this.x = (float) x;
+    this.y = (float) y;
+    this.z = (float) z;
   }
-  
+
   public Vector3f(float x, float y, float z) {
-    this.x = (float)x;
-    this.y = (float)y;
-    this.z = (float)z;
+    this.x = x;
+    this.y = y;
+    this.z = z;
   }
 
   public Vector3f(Vector3d other) {
     this(other.x, other.y, other.z);
   }
-  
+
   public Vector3f(Vector3f other) {
     this(other.x, other.y, other.z);
   }
@@ -37,11 +37,11 @@ public class Vector3f {
     this.y = y;
     this.z = z;
   }
-  
+
   public void set(Vector3f vec) {
     x = vec.x;
     y = vec.y;
-    z = vec.z;    
+    z = vec.z;
   }
 
   public void add(Vector3f vec) {
@@ -74,13 +74,13 @@ public class Vector3f {
     y *= scale;
     z *= scale;
   }
-  
+
   public double dot(Vector3f other) {
-    return x*other.x + y*other.y + z*other.z;
+    return x * other.x + y * other.y + z * other.z;
   }
 
   public double lengthSquared() {
-     return x*x + y*y + z*z;
+    return x * x + y * y + z * z;
   }
 
   public double length() {
